+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 12.0.0 (upcoming)
+
+* Pending changelog
+
 ## 11.0.0 (upcoming)
 
 **New features:**